Move changelog YAML serialization stack into Elastic.Documentation.Configuration
Centralizes all YAML file reading in Elastic.Documentation.Configuration.
The five serialization types (ChangelogConfigurationYaml, YamlLenientList,
YamlLenientListConverter, TypeEntryYamlConverter) and ChangelogConfigurationLoader
move from Elastic.Changelog into Configuration.Changelog/, registered in the
shared YamlStaticContext so the single-static-context-per-assembly constraint
is respected. Elastic.Changelog consumers drop the old using directives; no new
project references are needed since Changelog already references Configuration.
