+++ src/Elastic.Documentation.Configuration/Changelog/YamlLenientList.cs
@@ -2,14 +2,14 @@
 // Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
 // See the LICENSE file in the project root for more information
 
-namespace Elastic.Changelog.Serialization;
+namespace Elastic.Documentation.Configuration.Changelog;
 
 /// <summary>
 /// Wrapper type for YAML fields that can be specified as either a comma-separated string
 /// or a YAML list/sequence. Deserialized by <see cref="YamlLenientListConverter"/>.
 /// Uses mutable property for compatibility with YamlDotNet source generator.
 /// </summary>
-internal class YamlLenientList
+internal sealed class YamlLenientList
 {
 	public List<string>? Values { get; set; }
