+++ src/Elastic.Documentation.Configuration/Serialization/YamlStaticContext.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information
 
 using Elastic.Documentation.Configuration.Assembler;
+using Elastic.Documentation.Configuration.Changelog;
 using Elastic.Documentation.Configuration.Codex;
 using Elastic.Documentation.Configuration.LegacyUrlMappings;
 using Elastic.Documentation.Configuration.Products;
@@ -61,4 +62,18 @@ namespace Elastic.Documentation.Configuration.Serialization;
 [YamlSerializable(typeof(ChangelogConfigMinimalDto))]
 [YamlSerializable(typeof(RulesConfigMinimalDto))]
 [YamlSerializable(typeof(PublishRulesMinimalDto))]
+// Changelog YAML DTOs for CLI configuration (changelog.yml)
+[YamlSerializable(typeof(ChangelogConfigurationYaml))]
+[YamlSerializable(typeof(PivotConfigurationYaml))]
+[YamlSerializable(typeof(TypeEntryYaml))]
+[YamlSerializable(typeof(RulesConfigurationYaml))]
+[YamlSerializable(typeof(CreateRulesYaml))]
+[YamlSerializable(typeof(BundleRulesYaml))]
+[YamlSerializable(typeof(PublishRulesYaml))]
+[YamlSerializable(typeof(ProductsConfigYaml))]
+[YamlSerializable(typeof(DefaultProductYaml))]
+[YamlSerializable(typeof(BundleConfigurationYaml))]
+[YamlSerializable(typeof(BundleProfileYaml))]
+[YamlSerializable(typeof(ExtractConfigurationYaml))]
+[YamlSerializable(typeof(YamlLenientList))]
 public partial class YamlStaticContext;
