+++ src/services/Elastic.Changelog/Creation/ChangelogCreationService.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information
 
 using System.IO.Abstractions;
-using Elastic.Changelog.Configuration;
 using Elastic.Changelog.GitHub;
 using Elastic.Documentation;
 using Elastic.Documentation.Configuration;
