+++ src/services/Elastic.Changelog/Rendering/ChangelogRenderingService.cs
@@ -5,7 +5,6 @@
 using System.ComponentModel.DataAnnotations;
 using System.IO.Abstractions;
 using System.Text.Json.Serialization;
-using Elastic.Changelog.Configuration;
 using Elastic.Documentation;
 using Elastic.Documentation.Configuration;
 using Elastic.Documentation.Configuration.Changelog;
