+++ tests/Elastic.Changelog.Tests/Changelogs/ChangelogConfigurationTests.cs
@@ -3,8 +3,8 @@
 // See the LICENSE file in the project root for more information
 
 using AwesomeAssertions;
-using Elastic.Changelog.Configuration;
-using Elastic.Changelog.Serialization;
+using Elastic.Documentation.Configuration.Changelog;
+using Elastic.Documentation.Configuration.Changelog;
 using Elastic.Documentation;
 using Elastic.Documentation.Configuration;
 using Elastic.Documentation.Configuration.Changelog;
